+++ server/routes/brainImaging.ts
@@ -5,6 +5,7 @@ import type { CategoricalTW, BrainImagingRequest, BrainImagingResponse, FilesByC
 import { brainImagingPayload } from '#types'
 import { spawn } from 'child_process'
 import { getData } from '../src/termdb.matrix.js'
+import { isNumericTerm } from '@sjcrh/proteinpaint-shared/terms.js'
 
 /*
 given one or more samples, map the sample(s) to brain template and return the image
@@ -45,8 +46,8 @@ function init({ genomes }) {
 				index = query.t
 			}
 
-			const brainImage = await getBrainImage(query, genomes, plane, index)
-			res.send({ brainImage, plane } satisfies BrainImagingResponse)
+			const [brainImage, legend] = await getBrainImage(query, genomes, plane, index)
+			res.send({ brainImage, plane, legend })
 		} catch (e: any) {
 			console.log(e)
 			res.status(404).send('Sample brain image not found')
@@ -61,13 +62,10 @@ async function getBrainImage(query: BrainImagingRequest, genomes: any, plane: st
 	if (q[key].referenceFile && q[key].samples) {
 		const refFile = path.join(serverconfig.tpmasterdir, q[key].referenceFile)
 		const dirPath = path.join(serverconfig.tpmasterdir, q[key].samples)
-		const files = fs
-			.readdirSync(dirPath)
-			.filter(file => file.endsWith('.nii') && fs.statSync(path.join(dirPath, file)).isFile())
 
 		const terms: CategoricalTW[] = []
 		const divideByTW: CategoricalTW | undefined = query.divideByTW
-		const overlayTW = query.overlayTW
+		const overlayTW: CategoricalTW | undefined = query.overlayTW
 
 		if (divideByTW) terms.push(divideByTW)
 		if (overlayTW) terms.push(overlayTW)
@@ -82,15 +80,33 @@ async function getBrainImage(query: BrainImagingRequest, genomes: any, plane: st
 			const sampleId = ds.sampleName2Id.get(sampleName)
 			const sampleData = data.samples[sampleId]
 			const samplePath = path.join(dirPath, sampleName) + '.nii'
-			const divideCategory = divideByTW ? sampleData[divideByTW.$id!].value : 'default'
-			const overlayCategory = overlayTW ? sampleData[overlayTW.$id!].value : 'default'
+			let divideCategory = 'default'
+			let overlayCategory = 'default'
+			if (divideByTW) {
+				const value = sampleData[divideByTW.$id!]
+				if (value) divideCategory = divideByTW.term.values?.[value.key]?.label || value.key // for numeric terms key has the bin label
+			}
+			if (overlayTW) {
+				const value = sampleData[overlayTW.$id!]
+				if (value) overlayCategory = overlayTW.term.values?.[value.key]?.label || value.key
+			}
 			if (!divideByCat[divideCategory]) divideByCat[divideCategory] = {}
-			if (!divideByCat[divideCategory][overlayCategory])
-				divideByCat[divideCategory][overlayCategory] = {
-					samples: [],
-					color: overlayTW?.term?.values?.[overlayCategory]?.color || 'red'
+
+			if (!query.legendFilter?.includes(overlayCategory)) {
+				if (!divideByCat[divideCategory][overlayCategory]) {
+					let color = overlayTW?.term?.values?.[overlayCategory]?.color || 'red'
+
+					if (overlayTW && isNumericTerm(overlayTW.term)) {
+						const bins = data.refs.byTermId[overlayTW.$id!].bins
+						color = bins.find(b => b.label == overlayCategory).color
+					}
+					divideByCat[divideCategory][overlayCategory] = {
+						samples: [],
+						color
+					}
 				}
-			divideByCat[divideCategory][overlayCategory].samples.push(samplePath)
+				divideByCat[divideCategory][overlayCategory].samples.push(samplePath)
+			}
 		}
 		const lengths: number[] = []
 		for (const dcategory in divideByCat)
@@ -103,27 +119,31 @@ async function getBrainImage(query: BrainImagingRequest, genomes: any, plane: st
 		const maxLength = Math.max(...lengths)
 
 		const brainImageDict = {}
+		const legend = {}
 		for (const dcategory in divideByCat) {
 			let catNum = 0
 			const filesByCat = divideByCat[dcategory]
-			for (const category in filesByCat) catNum += filesByCat[category].samples.length
-			//if (samples.length < 1) continue
+			for (const category in filesByCat) {
+				if (filesByCat[category].samples.length < 1) continue
+				catNum += filesByCat[category].samples.length
+				if (!legend[category]) legend[category] = { color: filesByCat[category].color, maxLength }
+			}
 			const url = await generateBrainImage(refFile, plane, index, maxLength, JSON.stringify(filesByCat))
 			brainImageDict[dcategory] = { url, catNum }
 		}
 
-		return brainImageDict
+		if (query.legendFilter) {
+			for (const cat of query.legendFilter) {
+				legend[cat] = {
+					color: 'white',
+					maxLength,
+					crossedOut: true
+				}
+			}
+		}
+		return [brainImageDict, legend]
 	} else {
 		throw 'no reference or sample files'
-	}
-
-	function getFilesByCat(tw: CategoricalTW) {
-		const filesByCat: FilesByCategory = {}
-		for (const [key, value] of Object.entries(tw.term.values)) {
-			// TODO: make sure each term has a default color. buildTermDb assigns default color when not available
-			filesByCat[key] = { samples: [], color: value.color || 'red' }
-		}
-		return filesByCat
 	}
 }